myBar/requirements.txt
drf_yasg==1.20.0
Django==3.2.7
djangorestframework==3.12.4
pytest==5.4.0
pytest-cov==2.5.1
attrs==19.1.0
django-cors-headers==3.8.0
python-dateutil==2.2